Rename files and components related to cards
- Board-column changed to column
- New-board-column changed to new-column
- Column-card changed to card
- New-column-card changed to new-card
- Simplified folders
